# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/polyclinic/add": {
            "post": {
                "description": "Adds a polyclinic to the hospital",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Polyclinic"
                ],
                "summary": "Hastaneye poliklinik ekler",
                "parameters": [
                    {
                        "description": "Polyclinic info",
                        "name": "polyclinic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddHospitalPolyclinicRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.HospitalPolyclinicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/polyclinic/list": {
            "get": {
                "description": "Lists hospital's polyclinics with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Polyclinic"
                ],
                "summary": "Hastanenin polikliniklerini listeler (sayfalı)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HospitalPolyclinicListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/polyclinic/remove/{id}": {
            "delete": {
                "description": "Removes a polyclinic from the hospital",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Polyclinic"
                ],
                "summary": "Hastaneden poliklinik siler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hospital Polyclinic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/subuser": {
            "get": {
                "description": "Lists all subusers for the hospital",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubUser"
                ],
                "summary": "Alt kullanıcıları listeler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.SubUserResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new subuser (yetkili/çalışan)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubUser"
                ],
                "summary": "Alt kullanıcı ekler",
                "parameters": [
                    {
                        "description": "SubUser info",
                        "name": "subuser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.SubUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/subuser/{id}": {
            "put": {
                "description": "Updates a subuser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubUser"
                ],
                "summary": "Alt kullanıcıyı günceller",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SubUser ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SubUser info",
                        "name": "subuser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSubUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SubUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a subuser",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubUser"
                ],
                "summary": "Alt kullanıcıyı siler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SubUser ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddHospitalPolyclinicRequest": {
            "type": "object",
            "properties": {
                "polyclinic_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CreateSubUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "tc": {
                    "type": "string"
                }
            }
        },
...
        "dto.HospitalPolyclinicDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "personnel_groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PolyclinicPersonnelGroup"
                    }
                },
                "polyclinic_name": {
                    "type": "string"
                },
                "total_personnel": {
                    "type": "integer"
                }
            }
        },
        "dto.HospitalPolyclinicListResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "polyclinics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.HospitalPolyclinicDetail"
                    }
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.HospitalPolyclinicResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "polyclinic_id": {
                    "type": "integer"
                },
                "polyclinic_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.PolyclinicPersonnelGroup": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "group_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.SubUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "tc": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateSubUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "\"yetkili\" veya \"calisan\"",
                    "type": "string"
                },
                "tc": {
                    "type": "string"
                }
            }
        }
    }
}